requirements.txt
# Core scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI
anthropic==0.18.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
